refactor: extract parts of analytics control bar into its own components
