+++ src/components/analytics-control-bar-filter/README.md
@@ -1,0 +1,4 @@
+# AnalyticsControlBarFilter
+
+A filter is a underlined phrase that when clicked opens a popup underneath. It's used in the
+analytics control bar for different inputs.
