+++ src/components/analytics-control-bar-space-filter/README.md
@@ -1,0 +1,3 @@
+# AnalyticsControlBarSpaceFilter
+
+Creates a control bar filter in analytics used for scoping a query by the spaces within it.
